+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This file is used to list changes made in each version of the openvpn_okta cookb
 - Upgrade to the latest openvpn cookbook
 - Ensure gnupg is installed before trying to configure the repo
 - Ensure CA certs are installed and configured before configuring the repo
+- Ensure /etc/openvpn exists when needed
 
 ## 0.1.0 (2016-10-11)